feat(DriverDashboard): implement dashboard endpoint for driver with summary stats and new tour requests

+++ src/main/java/com/driverkonnect/backend/dto/response/driver/DriverActiveTourDto.java
@@ -21,7 +21,7 @@ public class DriverActiveTourDto {
     private Integer paxCount;
     private String travellerNationality;
     private String vehicleTypeName;
-    private Integer estimatedKm;
+    private BigDecimal estimatedKm;
     private String paymentTerm;
     private String specificRequirements;
     private String specialConcerns;
